backend_api/soar_engine/requirements.txt
fastapi
uvicorn
kafka-python
pyyaml
httpx